requirements.txt
# Google Drive Migration Helper - Requirements

# Google API 관련
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# 데이터 처리
pandas>=2.2.0
numpy>=1.26.0

# 진행률 표시
tqdm==4.66.1

# 로깅 및 설정
python-dotenv==1.0.0
pyyaml==6.0.1

# 파일 처리
pathlib2==2.3.7

# HTTP 요청
requests==2.31.0

# 날짜/시간 처리
python-dateutil==2.8.2

# JSON 처리
jsonschema==4.20.0

# 개발 및 테스트용
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Jupyter/Colab 지원
ipywidgets==8.1.1
jupyter==1.0.0
